Makefile test and code generator

+++ script/install_locally.sh
@@ -129,9 +129,10 @@ if [[ ${EL_MINIMAL_ADDONS} = "False" ]]; then
     printf "${EL_HOME}/addons/OCA_wms," >> ${EL_CONFIG_FILE}
     printf "${EL_HOME}/addons/Smile-SA_odoo_addons," >> ${EL_CONFIG_FILE}
     printf "${EL_HOME}/addons/TechnoLibre_odoo-code-generator," >> ${EL_CONFIG_FILE}
+    printf "${EL_HOME}/addons/TechnoLibre_odoo-code-generator-template," >> ${EL_CONFIG_FILE}
     printf "${EL_HOME}/addons/camptocamp_odoo-cloud-platform," >> ${EL_CONFIG_FILE}
     printf "${EL_HOME}/addons/dhongu_deltatech," >> ${EL_CONFIG_FILE}
-    printf "${EL_HOME}/addons/it-projects-llc_odoo-saas-tools," >> ${EL_CONFIG_FILE}
+    printf "${EL_HOME}/addons/it-projects-llc_saas-addons," >> ${EL_CONFIG_FILE}
     printf "${EL_HOME}/addons/itpp-labs_access-addons," >> ${EL_CONFIG_FILE}
     printf "${EL_HOME}/addons/itpp-labs_pos-addons," >> ${EL_CONFIG_FILE}
     printf "${EL_HOME}/addons/itpp-labs_website-addons," >> ${EL_CONFIG_FILE}
@@ -154,7 +155,7 @@ if [[ ${EL_MINIMAL_ADDONS} = "False" ]]; then
     printf "${EL_HOME}/doc/itpp-labs_odoo-port-docs," >> ${EL_CONFIG_FILE}
     printf "${EL_HOME}/doc/itpp-labs_odoo-test-docs," >> ${EL_CONFIG_FILE}
     printf "${EL_HOME}/doc/odoo_documentation-user," >> ${EL_CONFIG_FILE}
-    printf "${EL_HOME}/image_db," >> ${EL_CONFIG_FILE}
+    printf "${EL_HOME}/script/OCA_maintainer-tools," >> ${EL_CONFIG_FILE}
     printf "${EL_HOME}/script/OCA_odoo-module-migrator," >> ${EL_CONFIG_FILE}
 fi
 printf "\n" >> ${EL_CONFIG_FILE}
@@ -177,12 +178,20 @@ if [[ ! -d "./addons/addons" ]]; then
 fi
 
 PYENV_PATH=~/.pyenv
-PYENV_VERSION_PATH=${PYENV_PATH}/versions/3.7.7
+PYTHON_VERSION=3.7.7
+PYENV_VERSION_PATH=${PYENV_PATH}/versions/${PYTHON_VERSION}
 PYTHON_EXEC=${PYENV_VERSION_PATH}/bin/python
 POETRY_PATH=~/.poetry
 VENV_PATH=./.venv
+LOCAL_PYTHON_EXEC=${VENV_PATH}/bin/python
 VENV_REPO_PATH=${VENV_PATH}/repo
+VENV_MULTILINGUAL_MARKDOWN_PATH=${VENV_PATH}/multilang_md.py
 POETRY_VERSION=1.0.10
+
+echo "Python path version home"
+echo ${PYENV_VERSION_PATH}
+echo "Python path version local"
+echo ${LOCAL_PYTHON_EXEC}
 
 if [[ ! -d "${PYENV_PATH}" ]]; then
     echo -e "\n---- Installing pyenv in ${PYENV_PATH} ----"
@@ -195,28 +204,37 @@ eval "$(pyenv init -)"
 eval "$(pyenv virtualenv-init -)"
 
 if [[ ! -d "${PYENV_VERSION_PATH}" ]]; then
-    echo -e "\n---- Installing python 3.7.7 with pyenv in ${PYENV_VERSION_PATH} ----"
-    yes n|pyenv install 3.7.7
+    echo -e "\n---- Installing python ${PYTHON_VERSION} with pyenv in ${PYENV_VERSION_PATH} ----"
+    yes n|pyenv install ${PYTHON_VERSION}
+    if [[ $retVal -ne 0 ]]; then
+        echo "Error when installing pyenv"
+        exit 1
+    fi
 fi
 
-pyenv local 3.7.7
+pyenv local ${PYTHON_VERSION}
+
+if [[ ! -d ${VENV_PATH} ]]; then
+    echo -e "\n---- Create Virtual environment Python ----"
+    if [[ -e ${PYTHON_EXEC} ]]; then
+        ${PYTHON_EXEC} -m venv .venv
+        retVal=$?
+          if [[ $retVal -ne 0 ]]; then
+              echo "Virtual environment, error when creating .venv"
+              exit 1
+          fi
+    else
+        echo "Missing pyenv, please refer installation guide."
+        exit 1
+    fi
+fi
 
 if [[ ! -d "${POETRY_PATH}" ]]; then
     # Delete directory ~/.poetry and .venv to force update to new version
     echo -e "\n---- Installing poetry for reliable python package ----"
 #     curl -sSL https://raw.githubusercontent.com/python-poetry/poetry/master/get-poetry.py | ${PYTHON_EXEC}
     curl -fsS -o get-poetry.py https://raw.githubusercontent.com/python-poetry/poetry/master/get-poetry.py
-    python get-poetry.py -y --preview --version ${POETRY_VERSION}
-fi
-
-if [[ ! -d ${VENV_PATH} ]]; then
-    echo -e "\n---- Create Virtual environment Python ----"
-    if [[ -e ${PYTHON_EXEC} ]]; then
-        ${PYTHON_EXEC} -m venv .venv
-    else
-        echo "Missing pyenv, please refer installation guide."
-        exit 1
-    fi
+    ${LOCAL_PYTHON_EXEC} get-poetry.py -y --preview --version ${POETRY_VERSION}
 fi
 
 # Install git-repo if missing
@@ -229,11 +247,27 @@ if [[ ! -f ${VENV_REPO_PATH} ]]; then
     sed -i "1 i ${PYTHON_HASHBANG}" ${VENV_REPO_PATH}
 fi
 
+# Install Multilingual Markdown Generator if missing
+if [[ ! -f ${VENV_MULTILINGUAL_MARKDOWN_PATH} ]]; then
+    echo "\n---- Install Multilingual Markdown Generator ----"
+    curl https://raw.githubusercontent.com/ERPLibre/multilingual-markdown/master/multilang_md.py > ${VENV_MULTILINGUAL_MARKDOWN_PATH}
+    chmod +x ${VENV_MULTILINGUAL_MARKDOWN_PATH}
+    sed -i 1d ${VENV_MULTILINGUAL_MARKDOWN_PATH}
+    PYTHON_HASHBANG="#!./.venv/bin/python"
+    sed -i "1 i ${PYTHON_HASHBANG}" ${VENV_MULTILINGUAL_MARKDOWN_PATH}
+fi
+
 echo -e "\n---- Installing poetry dependency ----"
 ${VENV_PATH}/bin/pip install --upgrade pip
-#/home/"${USER}"/.poetry/bin/poetry env use ${PYTHON_EXEC}
-source $HOME/.poetry/env
-poetry install
+# Force python instead of changing env
+#/home/"${USER}"/.poetry/bin/poetry env use ${LOCAL_PYTHON_EXEC}
+# source $HOME/.poetry/env
+${LOCAL_PYTHON_EXEC} ~/.poetry/bin/poetry install
+retVal=$?
+if [[ $retVal -ne 0 ]]; then
+    echo "Poetry installation error."
+    exit 1
+fi
 # Delete artifacts created by pip, cause error in next "poetry install"
 rm -rf artifacts
